+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * improved tests related to time on page
 * fixed average time on page calculation
+* fixed events resetting time the session was last seen
 
 ## 6.5.1